requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.4.0
pandas>=1.5.0
python-dotenv>=0.21.0
click>=8.1.3

# LLM integrations
anthropic>=0.6.0
openai>=1.3.0
langchain>=0.0.200

# Web scraping extras
fake-useragent>=0.1.11
webdriver-manager>=3.8.5
html2text>=2020.1.16

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0